+++ kong/plugins/pre-function/_schema.lua
@@ -5,12 +5,26 @@ return function(plugin_name)
 
 
   local function validate_function(fun)
-    local _, err = loadstring(fun)
+    local func1, err = loadstring(fun)
     if err then
       return false, "Error parsing " .. plugin_name .. ": " .. err
     end
 
-    return true
+    local success, func2 = pcall(func1)
+
+    if not success or func2 == nil then
+      -- the code IS the handler function
+      return true
+    end
+
+    -- the code RETURNED the handler function
+    if type(func2) == "function" then
+      return true
+    end
+
+    -- the code returned something unknown
+    return false, "Bad return value from " .. plugin_name .. " function, " ..
+                  "expected function type, got " .. type(func2)
   end