+++ spec/02-schema_spec.lua
@@ -1,8 +1,10 @@
 local v = require("spec.helpers").validate_plugin_config_schema
 
-local mock_fn_one = 'print("hello world!")'
+local mock_fn_one = '("hello world!"):find("world")'
 local mock_fn_two = 'local x = 1'
+local mock_fn_three = 'local x = 1 return function() x = x + 1 end'
 local mock_fn_invalid = 'print('
+local mock_fn_invalid_return = 'return "hello-world"'
 
 for _, plugin_name in ipairs({ "pre-function", "post-function" }) do
 
@@ -21,6 +23,13 @@ for _, plugin_name in ipairs({ "pre-function", "post-function" }) do
       assert.falsy(err)
     end)
 
+    it("validates single function with upvalues", function()
+      local ok, err = v({ functions = { mock_fn_three } }, schema)
+
+      assert.truthy(ok)
+      assert.falsy(err)
+    end)
+
     it("validates multiple functions", function()
       local ok, err = v({ functions = { mock_fn_one, mock_fn_two } }, schema)
 
@@ -34,6 +43,13 @@ for _, plugin_name in ipairs({ "pre-function", "post-function" }) do
 
         assert.falsy(ok)
         assert.equals("Error parsing " .. plugin_name .. ": [string \"print(\"]:1: unexpected symbol near '<eof>'", err.config.functions[1])
+      end)
+
+      it("with an invalid return value", function()
+        local ok, err = v({ functions = { mock_fn_invalid_return } }, schema)
+
+        assert.falsy(ok)
+        assert.equals("Bad return value from " .. plugin_name .. " function, expected function type, got string", err.config.functions[1])
       end)
 
       it("with a valid and invalid function", function()